Refactor: declutter component
